Add files via upload
The Caesar cipher was supposedly developed by Julius Caesar as a way to secure his communications from his enemies. It is actually fairly simple. You simply shift each letter in the original text by a fixed amount and use the new letter in the encrypted message. For example, if you shift by three letters, your code would be:Original: abcdefghijklmnopqrstuvwxyzEncrypted: defghijklmnopqrstuvwxyzabcSo the text “caesar cipher server” would become “fdhvdu flskhu vhuyhu”. 

The server will accept an initial command of “ENCRYPT” or “DECRYPT” to tell it what to do with the incoming text. If the command sent over does not match either of these two commands, an error message is sent back and the client will not send over any text. If the command sent is correct, the server will respond with the word “OK”. The client is expecting this response and will not send the text if anything other than OK is received. Assuming the initial command was correct; the server will then accept lines (a string) of incoming text and will process them. The encrypted or decrypted text will be returned to the client.